pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "masterclass-mlops"
version = "0.1.0"
description = "Monitoring and observability masterclass repository"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "httpx>=0.28.0,<0.29.0",
  "prometheus-client>=0.21.0,<0.22.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "requests>=2.32.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "streamlit>=1.44.0,<2.0.0",
  "uvicorn>=0.34.0,<0.35.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.15.0,<2.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "ruff>=0.11.0,<0.12.0",
  "types-requests>=2.32.0,<3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/masterclass_mlops"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "N",
  "UP",
  "B",
  "C4",
  "SIM",
  "RUF",
]
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
plugins = []
strict = true
mypy_path = "src"
packages = ["masterclass_mlops"]
